Fix issue with not matching renders when using an underscore
This fixes #50. The regex being used to match render calls was incorrectly
not matching against variables with underscores in the name and so it was
using the wrong regex match data to populate the template dependency.
Because underscore is a valid character for a variable, this fixes the
ERBTracker to match it properly.